requirements.txt
pandas 
pyyaml
duckdb
streamlit
openpyxl